Questão 4

+++ utils.h
@@ -1,6 +1,6 @@
 #define utils_h
 
-int validaNaturais(float i){
+float validaNaturais(float i){
     if (i < 0){
         return 1;
     }else{
